requirements.txt
CyHunspell
